+++ include/CppUTest/TestHarness_c.h
@@ -192,39 +192,23 @@ extern "C"
 
 
 /* CHECKS that can be used from C code */
-extern void CHECK_EQUAL_C_BOOL_LOCATION(int expected, int actual,
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_INT_LOCATION(int expected, int actual,
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_UINT_LOCATION(unsigned int expected, unsigned int actual,
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_LONG_LOCATION(long expected, long actual,
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_ULONG_LOCATION(unsigned long expected, unsigned long actual,
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_LONGLONG_LOCATION(cpputest_longlong expected, cpputest_longlong actual,
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_ULONGLONG_LOCATION(cpputest_ulonglong expected, cpputest_ulonglong actual,
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_REAL_LOCATION(double expected, double actual,
-        double threshold, const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_CHAR_LOCATION(char expected, char actual,
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_UBYTE_LOCATION(unsigned char expected, unsigned char actual, 
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_SBYTE_LOCATION(signed char expected, signed char actual, 
-        const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_STRING_LOCATION(const char* expected,
-        const char* actual, const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_POINTER_LOCATION(const void* expected,
-        const void* actual, const char* text, const char* fileName, int lineNumber);
-extern void CHECK_EQUAL_C_BITS_LOCATION(unsigned int expected, unsigned int actual,
-        unsigned int mask, size_t size, const char* text, const char* fileName, int lineNumber);
-extern void FAIL_TEXT_C_LOCATION(const char* text, const char* fileName,
-        int lineNumber);
-extern void FAIL_C_LOCATION(const char* fileName, int lineNumber);
-extern void CHECK_C_LOCATION(int condition, const char* conditionString,
-        const char* text, const char* fileName, int lineNumber);
+extern void CHECK_EQUAL_C_BOOL_LOCATION(int expected, int actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_INT_LOCATION(int expected, int actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_UINT_LOCATION(unsigned int expected, unsigned int actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_LONG_LOCATION(long expected, long actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_ULONG_LOCATION(unsigned long expected, unsigned long actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_LONGLONG_LOCATION(cpputest_longlong expected, cpputest_longlong actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_ULONGLONG_LOCATION(cpputest_ulonglong expected, cpputest_ulonglong actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_REAL_LOCATION(double expected, double actual, double threshold, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_CHAR_LOCATION(char expected, char actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_UBYTE_LOCATION(unsigned char expected, unsigned char actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_SBYTE_LOCATION(signed char expected, signed char actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_STRING_LOCATION(const char* expected, const char* actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_POINTER_LOCATION(const void* expected, const void* actual, const char* text, const char* fileName, size_t lineNumber);
+extern void CHECK_EQUAL_C_BITS_LOCATION(unsigned int expected, unsigned int actual, unsigned int mask, size_t size, const char* text, const char* fileName, size_t lineNumber);
+extern void FAIL_TEXT_C_LOCATION(const char* text, const char* fileName, size_t lineNumber);
+extern void FAIL_C_LOCATION(const char* fileName, size_t lineNumber);
+extern void CHECK_C_LOCATION(int condition, const char* conditionString, const char* text, const char* fileName, size_t lineNumber);
 
 extern void* cpputest_malloc(size_t size);
 extern char* cpputest_strdup(const char* str);
@@ -233,14 +217,12 @@ extern void* cpputest_calloc(size_t num, size_t size);
 extern void* cpputest_realloc(void* ptr, size_t size);
 extern void  cpputest_free(void* buffer);
 
-extern void* cpputest_malloc_location(size_t size, const char* file, int line);
-extern char* cpputest_strdup_location(const char* str, const char* file, int line);
-extern char* cpputest_strndup_location(const char* str, size_t n, const char* file, int line);
-extern void* cpputest_calloc_location(size_t num, size_t size,
-        const char* file, int line);
-extern void* cpputest_realloc_location(void* memory, size_t size,
-        const char* file, int line);
-extern void cpputest_free_location(void* buffer, const char* file, int line);
+extern void* cpputest_malloc_location(size_t size, const char* file, size_t line);
+extern char* cpputest_strdup_location(const char* str, const char* file, size_t line);
+extern char* cpputest_strndup_location(const char* str, size_t n, const char* file, size_t line);
+extern void* cpputest_calloc_location(size_t num, size_t size, const char* file, size_t line);
+extern void* cpputest_realloc_location(void* memory, size_t size, const char* file, size_t line);
+extern void cpputest_free_location(void* buffer, const char* file, size_t line);
 
 void cpputest_malloc_set_out_of_memory(void);
 void cpputest_malloc_set_not_out_of_memory(void);
