+++ src/CppUTest/JUnitTestOutput.cpp
@@ -39,12 +39,12 @@ struct JUnitTestCaseResultNode
     }
 
     SimpleString name_;
-    long execTime_;
+    size_t execTime_;
     TestFailure* failure_;
     bool ignored_;
     SimpleString file_;
-    int lineNumber_;
-    int checkCount_;
+    size_t lineNumber_;
+    size_t checkCount_;
     JUnitTestCaseResultNode* next_;
 };
 
@@ -55,11 +55,11 @@ struct JUnitTestGroupResult
     {
     }
 
-    int testCount_;
-    int failureCount_;
-    int totalCheckCount_;
-    long startTime_;
-    long groupExecTime_;
+    size_t testCount_;
+    size_t failureCount_;
+    size_t totalCheckCount_;
+    size_t startTime_;
+    size_t groupExecTime_;
     SimpleString group_;
     JUnitTestCaseResultNode* head_;
     JUnitTestCaseResultNode* tail_;
@@ -110,8 +110,7 @@ void JUnitTestOutput::printCurrentGroupStarted(const UtestShell& /*test*/)
 
 void JUnitTestOutput::printCurrentTestEnded(const TestResult& result)
 {
-    impl_->results_.tail_->execTime_
-            = result.getCurrentTestTotalExecutionTime();
+    impl_->results_.tail_->execTime_ = result.getCurrentTestTotalExecutionTime();
     impl_->results_.tail_->checkCount_ = result.getCheckCount();
 }
 
@@ -130,7 +129,7 @@ void JUnitTestOutput::printCurrentTestStarted(const UtestShell& test)
 {
     impl_->results_.testCount_++;
     impl_->results_.group_ = test.getGroup();
-    impl_->results_.startTime_ = GetPlatformSpecificTimeInMillis();
+    impl_->results_.startTime_ = (size_t) GetPlatformSpecificTimeInMillis();
 
     if (impl_->results_.tail_ == NULLPTR) {
         impl_->results_.head_ = impl_->results_.tail_
@@ -189,9 +188,9 @@ void JUnitTestOutput::writeTestSuiteSummary()
             buf =
                     StringFromFormat(
                             "<testsuite errors=\"0\" failures=\"%d\" hostname=\"localhost\" name=\"%s\" tests=\"%d\" time=\"%d.%03d\" timestamp=\"%s\">\n",
-                            impl_->results_.failureCount_,
+                            (int)impl_->results_.failureCount_,
                             impl_->results_.group_.asCharString(),
-                            impl_->results_.testCount_,
+                            (int) impl_->results_.testCount_,
                             (int) (impl_->results_.groupExecTime_ / 1000), (int) (impl_->results_.groupExecTime_ % 1000),
                             GetPlatformSpecificTimeString());
     writeToFile(buf.asCharString());
@@ -225,10 +224,10 @@ void JUnitTestOutput::writeTestCases()
                 impl_->package_.isEmpty() ? "" : ".",
                 impl_->results_.group_.asCharString(),
                 cur->name_.asCharString(),
-                cur->checkCount_ - impl_->results_.totalCheckCount_,
+                (int) (cur->checkCount_ - impl_->results_.totalCheckCount_),
                 (int) (cur->execTime_ / 1000), (int)(cur->execTime_ % 1000),
                 cur->file_.asCharString(),
-                cur->lineNumber_);
+                (int) cur->lineNumber_);
         writeToFile(buf.asCharString());
 
         impl_->results_.totalCheckCount_ = cur->checkCount_;
@@ -249,7 +248,7 @@ void JUnitTestOutput::writeFailure(JUnitTestCaseResultNode* node)
     SimpleString buf = StringFromFormat(
             "<failure message=\"%s:%d: %s\" type=\"AssertionFailedError\">\n",
             node->failure_->getFileName().asCharString(),
-            node->failure_->getFailureLineNumber(),
+            (int) node->failure_->getFailureLineNumber(),
             encodeXmlText(node->failure_->getMessage()).asCharString());
     writeToFile(buf.asCharString());
     writeToFile("</failure>\n");
