+++ src/CppUTestExt/MemoryReportAllocator.cpp
@@ -77,7 +77,7 @@ void MemoryReportAllocator::setFormatter(MemoryReportFormatter* formatter)
     formatter_ = formatter;
 }
 
-char* MemoryReportAllocator::alloc_memory(size_t size, const char* file, int line)
+char* MemoryReportAllocator::alloc_memory(size_t size, const char* file, size_t line)
 {
     char* memory = realAllocator_->alloc_memory(size, file, line);
     if (result_ && formatter_)
@@ -85,7 +85,7 @@ char* MemoryReportAllocator::alloc_memory(size_t size, const char* file, int lin
     return memory;
 }
 
-void MemoryReportAllocator::free_memory(char* memory, const char* file, int line)
+void MemoryReportAllocator::free_memory(char* memory, const char* file, size_t line)
 {
     realAllocator_->free_memory(memory, file, line);
     if (result_ && formatter_)
